requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
aws-requests-auth==0.4.3
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.39.3
botocore==1.39.17
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
gevent==25.5.1
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.1
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
opentelemetry-api==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
readabilipy==0.3.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.36.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.3
strands-agents==1.5.0
strands-agents-tools==0.2.4
sympy==1.14.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid==1.30
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2